requirements.txt
odc-stats  # Installing here so maybe the plugin will register properly
gdal
rasterio
# # Pin some packages
numpy==1.21.5
xarray==2023.1.0
# This is datacube_compute which is a custom build
https://github.com/auspatious/datacube-compute/releases/download/0.0.6/datacube_compute-0.0.6-cp310-cp310-linux_x86_64.whl

--no-binary gdal
--no-binary rasterio
